+++ Geometry2d/include/Geometry2d/CompositeShape.hpp
