+++ Geometry2d/include/Geometry2d/Line.hpp
