+++ Geometry2d/include/Geometry2d/Polygon.hpp
