+++ Geometry2d/include/Geometry2d/Rect.hpp
