+++ Geometry2d/include/Geometry2d/Segment.hpp
