+++ Geometry2d/include/Geometry2d/Shape.hpp
