+++ Geometry2d/include/Geometry2d/ShapeSet.hpp
