+++ Geometry2d/src/CompositeShape.cpp
@@ -1,4 +1,4 @@
-#include "geometry2d/CompositeShape.hpp"
+#include <Geometry2d/CompositeShape.hpp>
 
 namespace Geometry2d {
