+++ Geometry2d/src/Polygon.cpp
@@ -1,4 +1,4 @@
-#include "geometry2d/Polygon.hpp"
+#include <Geometry2d/Polygon.hpp>
 
 #include <rj_constants/constants.hpp>
