+++ Geometry2d/src/Rect.cpp
@@ -1,6 +1,6 @@
-#include "geometry2d/Rect.hpp"
-#include "geometry2d/Point.hpp"
-#include "geometry2d/Segment.hpp"
+#include <Geometry2d/Rect.hpp>
+#include <Geometry2d/Point.hpp>
+#include <Geometry2d/Segment.hpp>
 #include <rj_constants/constants.hpp>
 
 using namespace std;
