+++ Geometry2d/src/Segment.cpp
@@ -1,7 +1,7 @@
 
-#include "geometry2d/Segment.hpp"
+#include <Geometry2d/Segment.hpp>
 
-#include "geometry2d/Util.hpp"
+#include <Geometry2d/Util.hpp>
 
 using namespace std;
